Setup database migration scripts using alembic. see #81
Database migrations can only be upgrades if used through the pacsanini
API. There shouldn't be a need for users to perform other types of
activities.
The tricky part was to implement a mechanism allowing for the creation/
dropping of views. Fortunately, alembic has a good cookbook for this
(see the link in the code).

Database initialization also now uses alembic. This is so that the
alembic_version table can be written on initialization. Code changes had
to occur (refactoring only) to adapt this new method for when the db is
dynamically created during c-find operations.

Documentation was updated to introduce what the command can do.

Signed-off-by: Aurélien Chick <aurelien.chick@gmail.com>